Create OrderTest
This code completes the `OrderTest` class and adds a test for the `get_order_history` method. The `testAddAndGetAllOrders` and `testGetOrderHistory` methods test the functionality of adding orders and retrieving them, both for all orders and for specific email addresses.

To run the test, execute the following command in the terminal:


If all tests pass, you should see output indicating the success of each test case.